+++ core/templates/taskInfo.html
@@ -386,10 +386,15 @@
   <div class="columns small-5 medium-4 large-3 align-right">
   {% with 'done finished failed ready broken aborted defined' as finalstatelist %}
       {% if not task.status in finalstatelist %}
-          <a id="button-task-action-finish-panda" class="bp-button kill">Finish</a>
-          <a id="button-task-action-abort-panda" class="bp-button kill">Abort</a>
-      {% endif %}
+          {% if 'indigoiam' in authtype %}
+              <a id="button-task-action-finish-panda" class="bp-button kill">Finish (PanDA client)</a>
+              <a id="button-task-action-abort-panda" class="bp-button kill">Abort (PanDA client)</a>
+          {% else %}
+              <a id="button-task-action-finish-prodsys" class='bp-button kill'>Finish</a>
+              <a id="button-task-action-abort-prodsys" class='bp-button kill'>Abort</a>
+          {% endif %}
 
+      {% endif %}
   {% endwith %}
   </div>
   </div>
@@ -810,6 +815,8 @@ document.getElementById("button-task-plots").addEventListener("click", togglePlo
 
 if (document.getElementById("button-task-action-finish-panda")) {document.getElementById("button-task-action-finish-panda").addEventListener("click", operationtask_token.bind(null, jeditaskid, 'finishtask'));}
 if (document.getElementById("button-task-action-abort-panda")) {document.getElementById("button-task-action-abort-panda").addEventListener("click", operationtask_token.bind(null, jeditaskid, 'killtask' ));}
+if (document.getElementById("button-task-action-finish-prodsys")) {document.getElementById("button-task-action-finish-prodsys").addEventListener("click", killtasks.bind(null, jeditaskid, 0));}
+if (document.getElementById("button-task-action-abort-prodsys")) {document.getElementById("button-task-action-abort-prodsys").addEventListener("click", killtasks.bind(null, jeditaskid, 1));}
 
 const task_age = {{ task.age|safe }};
 const task_rating = {{ task.rating|safe }};
@@ -1526,7 +1533,7 @@ function operationtask_token(taskID, action) {
     if (confirm(message)) {
 
         $.ajax({
-            url: '/panda_client/',
+            url: 'panda_client/',
             data: {taskID: taskID, action: action},
             async: false,
             type: "POST",
